requirements.txt
torch
diffusers
transformers
accelerate
safetensors
gradio
Pillow
numpy
python-dotenv
